backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas>=2.2.1
google-generativeai>=0.3.2
python-multipart>=0.0.21
python-dotenv==1.0.0
yfinance==0.2.32
pillow>=10.2.0
pydantic>=2.8.0